Update CI and remove deprecated PHP resource cleanup
- Add PHP 8.5 to CI matrix and pest ^4.0 support
- Update phpunit.xml.dist to PHPUnit 12.5 schema
- Cast floor() to int in ColorSpaceConverter to fix deprecation
- Remove finfo_close() calls (auto-managed since PHP 8.0)
- Remove imagedestroy() calls (GdImage objects auto-cleanup since PHP 8.0)

+++ src/ColorSpaceConverter.php
@@ -241,7 +241,7 @@ class ColorSpaceConverter
         }
 
         $h = $h * 6;
-        $i = floor($h);
+        $i = (int) floor($h);
         $f = $h - $i;
         $p = $v * (1 - $s);
         $q = $v * (1 - $s * $f);
